Cleaned up WET Manage code
